requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
# Forzamos una versión reciente para evitar el error 404 de los modelos nuevos
google-generativeai>=0.5.0
openai>=1.0.0
anthropic>=0.20.0